Add files via upload
The objective of this project was to explore and analyze the primary asset allocation strategies in depth, highlighting their differences and emphasizing their practical implementation. This assignment was designed as a take-home exam, meaning some aspects may have been tailored for educational purposes. Nonetheless, the focus was on strengthening Python programming skills and gaining a clearer understanding of the advantages and limitations of these methodologies.